// src/helper/types.ts
export interface WalletHistoryProps {
  _id: string;
  user_id: string;
  currency: string;
  currency_id: string;
  chain: string;
  short_name: string;
  deposit_slip: string;
  transaction_number: number;
  description: string;
  amount: number;
  transaction_type: string;
  fee: number;
  status: string;
  from_address: string;
  to_address: string;
  createdAt: string;
  updatedAt: string;
  __v: number;
}
export interface TradeHistoryProps {
  _id: string;
  user_id: string;
  order_id: string;
  currency: string;
  currency_id: string;
  quantity: number;
  price: number;
  amount: number;
  side: string;
  transaction_type: string;
  order_type: string;
  fee_type: string;
  fee: number;
  tds: number;
  createdAt: string;
  updatedAt: string;
  __v: number;
}